fix: REST API

+++ controller/src/main/java/ru/kramskoi/controller/OwnerController.java
@@ -1,7 +1,8 @@
 package ru.kramskoi.controller;
 
-
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.kramskoi.dto.OwnerDTO;
@@ -11,37 +12,49 @@ import ru.kramskoi.service.OwnerService;
 import javax.validation.Valid;
 
 @RestController
-@RequestMapping("/owner")
+@RequestMapping("/owners")
 @Validated
-
 public class OwnerController {
+
+    @Autowired
     private final OwnerService ownerService;
 
     public OwnerController(OwnerService ownerService) {
         this.ownerService = ownerService;
     }
 
-    @GetMapping("/getOwnerById")
-    @ResponseStatus(HttpStatus.OK)
-    public OwnerDTO getOwnerById(@RequestParam("id") Long id) {
-        return ownerService.getOwnerByID(id);
+    @GetMapping("/{id}")
+    public ResponseEntity<OwnerDTO> getOwnerById(@PathVariable("id") Long id) {
+        OwnerDTO ownerDTO = ownerService.getOwnerByID(id);
+        if (ownerDTO == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(ownerDTO, HttpStatus.OK);
     }
 
-    @PutMapping("/updateOwner")
-    @ResponseStatus(HttpStatus.OK)
-    public void updateOwner(@Valid @RequestBody OwnerDTO ownerDTO) {
+    @PutMapping("/{id}")
+    public ResponseEntity<Void> updateOwner(@PathVariable("id") Long id, @Valid @RequestBody OwnerDTO ownerDTO) {
+        OwnerDTO existingOwnerDTO = ownerService.getOwnerByID(id);
+        if (existingOwnerDTO == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         ownerService.updateOwner(OwnerMapper.fromDTOToOwner(ownerDTO));
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PostMapping("/addOwner")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void addOwner(@Valid @RequestBody OwnerDTO ownerDTO) {
+    @PostMapping
+    public ResponseEntity<Void> addOwner(@Valid @RequestBody OwnerDTO ownerDTO) {
         ownerService.addOwner(OwnerMapper.fromDTOToOwner(ownerDTO));
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    @DeleteMapping("/deleteOwner")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteOwner(@RequestParam("id") Long id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteOwner(@PathVariable("id") Long id) {
+        OwnerDTO ownerDTO = ownerService.getOwnerByID(id);
+        if (ownerDTO == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         ownerService.deleteOwner(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
